Update .gitignore, add highly active user analysis
For the Newcomer Tasks analysis, we want to know whether we've
meaningfully changed the proportion of newcomers who go on to
become "Highly active users". This notebook did a quick lit
review to identify some papers that talk about that, and how
they go about measuring it. Then we do a quick dive into
data from 2019 on newcomer activity.